chore: 🚚 rename using kebab-case
